Fix cgo filtering
This changes our filtering code to understand the implied build tags caused by
import "C" lines in go files that build.Context.Match does not handle.
It also changes our cgo processing code to share the same logic.
