+++ tests/cgo_pure/cgo_no_tag.go
@@ -1,0 +1,8 @@
+package cgo_pure
+
+/*
+const int value;
+*/
+import "C"
+
+var AnotherValue = int(C.value)
